log_in/requirements.txt
asgiref==3.11.1
crispy-bootstrap5==2025.6
Django==4.2.28
django-crispy-forms==2.5
django-formtools==2.5.1
django-otp==1.7.0
django-phonenumber-field==8.3.0
django-simple-captcha==0.6.3
django-two-factor-auth==1.18.1
djangorestframework==3.16.1
python-decouple==3.8
qrcode==8.2
sqlparse==0.5.5
typing_extensions==4.15.0
